optimize js structure
