+++ assets/js/lib.js
@@ -1,0 +1,2 @@
+import './include/jquery.js'
+import './include/jquery.fitvids.js'
