+++ default.hbs
@@ -59,9 +59,7 @@
 {{ghost_foot}}
 
 {{! JavaScripts }}
-<script type="text/javascript" src="{{asset "js/jquery.min.js"}}"></script>
-<script type="text/javascript" src="{{asset "js/jquery.fitvids.min.js"}}"></script>
-<script type="text/javascript" src="{{asset "js/search.min.js"}}"></script>
+<script type="text/javascript" src="{{asset "js/lib.min.js"}}"></script>
 <script type="text/javascript" src="{{asset "js/index.min.js"}}"></script>
 
 </body>
