+++ package.json
@@ -24,6 +24,7 @@
     "del": "^5.1.0",
     "gulp": "^4.0.2",
     "gulp-clean-css": "^4.3.0",
+    "gulp-js-import-file": "^1.0.11",
     "gulp-minify": "^3.1.0",
     "gulp-sass": "^4.1.0",
     "gulp-zip": "^5.0.2",
